Reorganize files and use existing tests for longest increasing subsequence
